dashboard/requirements.txt
streamlit
requests
pandas
plotly-express
graphviz